+++ __tests__/lib/dev-environment/dev-environment-core.js
@@ -148,6 +148,8 @@ describe( 'lib/dev-environment/dev-environment-core', () => {
 						branch: 'dev',
 						isMultisite: true,
 						primaryDomain: '',
+						php: '',
+						wordpress: '',
 					},
 				},
 			},
@@ -193,6 +195,18 @@ describe( 'lib/dev-environment/dev-environment-core', () => {
 							primaryDomain: {
 								name: 'test.develop.com',
 							},
+							softwareSettings: {
+								php: {
+									current: {
+										version: '8.1',
+									},
+								},
+								wordpress: {
+									current: {
+										version: '6.2',
+									},
+								},
+							},
 						},
 						{
 							name: 'prodName',
@@ -212,6 +226,8 @@ describe( 'lib/dev-environment/dev-environment-core', () => {
 						branch: 'dev',
 						isMultisite: true,
 						primaryDomain: 'test.develop.com',
+						php: '8.1',
+						wordpress: '6.2',
 					},
 				},
 			},
@@ -262,12 +278,13 @@ describe( 'lib/dev-environment/dev-environment-core', () => {
 			expect( searchAndReplace ).not.toHaveBeenCalled();
 
 			await expect( promise ).rejects.toEqual(
-				new Error( 'The provided file does not exist or it is not valid (see "--help" for examples)' )
+				new Error( 'The provided file undefined does not exist or it is not valid (see "--help" for examples)' )
 			);
 		} );
 
 		it( 'should resolve the path and replace it with /user', async () => {
 			fs.existsSync.mockReturnValue( true );
+			fs.lstatSync.mockReturnValue( { isDirectory: () => false } );
 			const resolvedPath = `${ os.homedir() }/testfile.sql`;
 			resolvePath.mockReturnValue( resolvedPath );