+++ src/lib/dev-environment/dev-environment-cli.js
@@ -38,7 +38,10 @@ const DEFAULT_SLUG = 'vip-local';
 
 export async function handleCLIException( exception: Error, trackKey?: string, trackBaseInfo?: any = {} ) {
 	const errorPrefix = chalk.red( 'Error:' );
-	if ( DEV_ENVIRONMENT_NOT_FOUND === exception.message ) {
+	if ( exception instanceof UserError ) {
+		// User errors are handled in global error handler
+		throw exception;
+	} else if ( DEV_ENVIRONMENT_NOT_FOUND === exception.message ) {
 		const createCommand = chalk.bold( DEV_ENVIRONMENT_FULL_COMMAND + ' create' );
 
 		const message = `Environment doesn't exist.\n\n\nTo create a new environment run:\n\n${ createCommand }\n`;
@@ -69,12 +72,37 @@ export async function handleCLIException( exception: Error, trackKey?: string, t
 	}
 }
 
-export const validateDependencies = async () => {
+const verifyDNSResolution = ( slug: string ) => {
+	const dns = require( 'dns' );
+	const expectedIP = '127.0.0.1';
+	const testDomain = `${ slug }.vipdev.lndo.site`;
+	const advice = `Please add following line to hosts file on your system:\n${ expectedIP } ${ testDomain }`;
+
+	debug( `Verifying DNS resolution for ${ testDomain }` );
+	return new Promise( ( resolve, reject ) => {
+		dns.lookup( testDomain, ( error, address ) => {
+			debug( `Got DNS response ${ address }` );
+
+			if ( error ) {
+				reject( new UserError( `DNS resolution for ${ testDomain } failed. ${ advice }` ) );
+			}
+
+			if ( address !== expectedIP ) {
+				reject( new UserError( `DNS resolution for ${ testDomain } returned unexpected IP ${ address }. Expected value is ${ expectedIP }. ${ advice }` ) );
+			}
+
+			resolve();
+		} );
+	} );
+};
+
+export const validateDependencies = async ( slug: string ) => {
 	try {
 		await validateDockerInstalled();
 	} catch ( exception ) {
-		exit.withError( exception.message );
+		throw new UserError( exception.message );
 	}
+	await verifyDNSResolution( slug );
 };
 
 export function getEnvironmentName( options: EnvironmentNameOptions ): string {
@@ -132,6 +160,8 @@ export function getOptionsFromAppInfo( appInfo: AppInfo ): InstanceOptions {
 		title: appInfo.environment?.name || appInfo.name || '',
 		multisite: !! appInfo?.environment?.isMultisite,
 		mediaRedirectDomain: appInfo.environment?.primaryDomain,
+		php: appInfo.environment?.php || '',
+		wordpress: appInfo.environment?.wordpress || '',
 	};
 }